# src/main/resources/swagger.yaml
openapi: 3.0.1
info:
  title: 12 Kingdoms Backend
  version: V2
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /players:
    get:
      tags:
        - player-controller-impl
      summary: Retrieve player details
      operationId: playerDetails
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerModel'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
  /players/list:
    get:
      tags:
        - player-controller-impl
      summary: List players
      operationId: playerList
      parameters:
        - name: currentPage
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 5
        - name: sortDirection
          in: query
          required: false
          schema:
            type: string
        - name: sortField
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerListResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
...
  /players/{id}:
    delete:
      tags:
        - player-controller-impl
      summary: Delete a player
      operationId: deletePlayer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    ErrorDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    ErrorResponse:
      type: object
      properties:
        errorList:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDto'
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    WoundsModel:
      required:
        - description
        - diceRange
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        diceRange:
          type: string
        description:
          type: string
    AdvantageModel:
      required:
        - description
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    AptitudeModel:
      required:
        - description
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    BackgroundModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        physicalPoints:
          type: integer
          format: int32
        mentalPoints:
          type: integer
          format: int32
        advantage:
          $ref: '#/components/schemas/AdvantageModel'
    JobModel:
      required:
        - description
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        mainAttribute:
          type: string
          enum:
            - INTELLIGENCE
            - CUNNING
            - CELERITY
            - TENACITY
        physicalPoints:
          type: integer
          format: int32
        mentalPoints:
          type: integer
          format: int32
        physicalPerLevel:
          type: integer
          format: int32
        mentalPerLevel:
          type: integer
          format: int32
        skills:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/SkillModel'
        aptitudes:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/AptitudeModel'
    LineageModel:
      required:
        - description
        - language
        - maturity
        - negativeTraitDescription
        - negativeTraitName
        - positiveTraitDescription
        - positiveTraitName
        - size
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        size:
          type: string
        language:
          type: string
        maturity:
          type: string
        positiveTraitName:
          type: string
        positiveTraitDescription:
          type: string
        negativeTraitName:
          type: string
        negativeTraitDescription:
          type: string
    SheetModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        intelligence:
          type: string
          enum:
            - D4
            - D6
            - D8
            - D12
            - D20
        cunning:
          type: string
          enum:
            - D4
            - D6
            - D8
            - D12
            - D20
        tenacity:
          type: string
          enum:
            - D4
            - D6
            - D8
            - D12
            - D20
        celerity:
          type: string
          enum:
            - D4
            - D6
            - D8
            - D12
            - D20
        name:
          type: string
        level:
          type: integer
          format: int32
        mentalCurrent:
          type: integer
          format: int32
        mentalTotal:
          type: integer
          format: int32
        physicalCurrent:
          type: integer
          format: int32
        physicalTotal:
          type: integer
          format: int32
        heroismCurrent:
          type: integer
          format: int32
        heroismTotal:
          type: integer
          format: int32
        bond:
          type: string
        motivation:
          type: string
        notes:
          type: string
        deathRollBody:
          type: string
          enum:
            - UNCHECKED
            - SUCCESS
            - FAILURE
        deathRollMind:
          type: string
          enum:
            - UNCHECKED
            - SUCCESS
            - FAILURE
        deathRollSpirit:
          type: string
          enum:
            - UNCHECKED
            - SUCCESS
            - FAILURE
        lineage:
          $ref: '#/components/schemas/LineageModel'
        background:
          $ref: '#/components/schemas/BackgroundModel'
        job:
          $ref: '#/components/schemas/JobModel'
        wound:
          $ref: '#/components/schemas/WoundsModel'
        aptitudes:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/AptitudeModel'
        active:
          type: boolean
    SkillModel:
      required:
        - description
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        cost:
          type: integer
          format: int32
        energyType:
          type: string
          enum:
            - PHYSICAL
            - MENTAL
        skillType:
          type: string
          enum:
            - INTERPRETATION
            - COMBAT
            - BASIC
        range:
          type: integer
          format: int32
        skillLevel:
          type: integer
          format: int32
...
    CampaignModel:
      required:
        - active
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        player:
          $ref: '#/components/schemas/PlayerModel'
        sheets:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/SheetModel'
        active:
          type: boolean
    PlayerModel:
      required:
        - email
        - firstName
        - lastName
        - username
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        roles:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/RoleModel'
        sheets:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/SheetModel'
        campaigns:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/CampaignModel'
        active:
          type: boolean
    RoleModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          enum:
            - ROLE_USER
            - ROLE_GM
            - ROLE_ADMIN
    BasicPlayerDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        active:
          type: boolean
    PlayerListResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        currentPage:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        list:
          type: array
          items:
            $ref: '#/components/schemas/BasicPlayerDto'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT